chore: move index from lib to widgets
