+++ src/widgets/index.ts
@@ -37,4 +37,4 @@ export {
 export {
   default as queryRuleContext,
 } from './query-rule-context/query-rule-context';
-export { default as index } from '../lib/index';
+export { default as index } from './index/index';
